# core/docs/swagger.yaml
definitions:
...
  internal_api_handlers.PolicyCreateRequest:
    properties:
      action:
        type: string
      dst-addr:
        type: string
      dst-zone:
        type: string
      log:
        type: boolean
      services:
        items:
          type: string
        type: array
      src-addr:
        type: string
      src-zone:
        type: string
    type: object
  internal_api_handlers.PolicyEditRequest:
    properties:
      action:
        type: string
      dst-addr:
        type: string
      dst-zone:
        type: string
      log:
        type: boolean
      services:
        items:
          type: string
        type: array
      src-addr:
        type: string
      src-zone:
        type: string
    type: object
  internal_api_handlers.PolicyMoveRequest:
    properties:
      position:
        type: string
      reference_id:
        type: integer
    type: object
...
  net.IPNet:
    properties:
      ip:
        description: network number
        items:
          format: int32
          type: integer
        type: array
      mask:
        description: network mask
        items:
          format: int32
          type: integer
        type: array
    type: object
  sentinelos_core_internal_model.Action:
    enum:
    - allow
    - deny
    type: string
    x-enum-varnames:
    - Allow
    - Deny
  sentinelos_core_internal_model.Address:
    properties:
      ips:
        items:
          $ref: '#/definitions/net.IPNet'
        type: array
      name:
        type: string
    type: object
...
  sentinelos_core_internal_model.Policy:
    properties:
      action:
        $ref: '#/definitions/sentinelos_core_internal_model.Action'
      dstAddr:
        $ref: '#/definitions/sentinelos_core_internal_model.Address'
      dstZone:
        $ref: '#/definitions/sentinelos_core_internal_model.Zone'
      id:
        type: integer
      log:
        type: boolean
      services:
        items:
          $ref: '#/definitions/sentinelos_core_internal_model.Service'
        type: array
      srcAddr:
        $ref: '#/definitions/sentinelos_core_internal_model.Address'
      srcZone:
        $ref: '#/definitions/sentinelos_core_internal_model.Zone'
    type: object
  sentinelos_core_internal_model.Protocol:
    enum:
    - tcp
    - udp
    type: string
    x-enum-varnames:
    - TCP
    - UDP
...
  sentinelos_core_internal_model.Service:
    properties:
      name:
        type: string
      ports:
        items:
          type: integer
        type: array
      protocol:
        $ref: '#/definitions/sentinelos_core_internal_model.Protocol'
    type: object
...
  sentinelos_core_internal_model.Zone:
    properties:
      color:
        type: string
      interfaces:
        items:
          type: string
        type: array
      name:
        type: string
      networks:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/sentinelos_core_internal_model.ZoneType'
    type: object
  sentinelos_core_internal_model.ZoneType:
    enum:
    - l2
    - l3
    type: string
    x-enum-varnames:
    - ZoneL2
    - ZoneL3
...
  sentinelos_core_pkg_utils.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: Next-Generation Firewall Management API.
  title: SentinelOS API
  version: "1.0"
paths:
...
  /api/policies:
    get:
      description: Get a list of all policies.
      produces:
      - application/json
      responses:
        "200":
          description: List of policies
          schema:
            items:
              $ref: '#/definitions/sentinelos_core_internal_model.Policy'
            type: array
        "500":
          description: ERR_SYS_4001 Internal server error
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: List Policies
      tags:
      - security
      - policies
    post:
      consumes:
      - application/json
      description: Create a new security policy.
      parameters:
      - description: Policy details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.PolicyCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: policy created in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON or references unknown entity
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Create Policy
      tags:
      - security
      - policies
  /api/policies/{id}:
    delete:
      description: Delete an existing security policy.
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: policy deleted from candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "404":
          description: ERR_NET_1003 Resource not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Delete Policy
      tags:
      - security
      - policies
    put:
      consumes:
      - application/json
      description: Update an existing security policy.
      parameters:
      - description: Policy ID
        in: path
        name: id
        required: true
        type: integer
      - description: Policy details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.PolicyEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: policy updated in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON or ID
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "404":
          description: ERR_NET_1003 Resource not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Edit Policy
      tags:
      - security
      - policies
  /api/policies/{id}/move:
    post:
      consumes:
      - application/json
      description: Move a policy relative to another policy (top, bottom, before,
        after).
      parameters:
      - description: Policy ID to move
        in: path
        name: id
        required: true
        type: integer
      - description: Move details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.PolicyMoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: policy moved successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON or Position
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "404":
          description: ERR_NET_1003 Policy not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Move Policy
      tags:
      - security
      - policies
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"